Migrate BlacklistInfoEventHandler to BlacklistEventHandler
